% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCosine.R
\name{clusterCosine}
\alias{clusterCosine}
\title{Cosine simillarity clustering}
\usage{
clusterCosine(ge, k)
}
\arguments{
\item{ge}{matrix, given gene expression dataset}

\item{k}{numeric, number of clusters}
}
\value{
matrix, clustered dataset
}
\description{
Clusters given dataset using cosine simillarity as distance
Uses Kmeans from amap
}

