% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countInside.R
\name{countInside}
\alias{countInside}
\title{Check how many points are inside of given simplex}
\usage{
countInside(V, P)
}
\arguments{
\item{V}{matrix, given simplex, n \times (n + 1) matrix describing (n + 1) points in n-dimensional space}

\item{P}{matrix, given points to check, n \times N matrix describing N points in n-dimensional space}
}
\value{
logical vector wether point is inside of simplex
}
\description{
Barycentric transformation can be applied to set of points
to check which of points are lie inside fiven simplex
}
