% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseLinearFit.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolution by it self}
\usage{
deconvolve(mixedData, gg, coefMatrix, coef, dims)
}
\value{
list containing basis, proportions and reconstructed results
}
\description{
Deconvolution by it self
}
