% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_utils.R
\name{dim_error}
\alias{dim_error}
\title{Out of space error}
\usage{
dim_error(m, m_red)
}
\arguments{
\item{m}{given matrix M}

\item{m_red}{approximation of M of lower rank}
}
\value{
sorted vector of errors
}
\description{
Calculates an error between given matrix M and its approximation matrix M_red of lower rank for every gene
}
