% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_utils.R
\name{dimensionality_reconstruction}
\alias{dimensionality_reconstruction}
\title{Dimensionality reconstruction}
\usage{
dimensionality_reconstruction(m, dims)
}
\arguments{
\item{m}{given matrix to be approximated}

\item{dims}{dimensionality of return matrix}
}
\value{
approximation of m with matrix of rank dims
}
\description{
Returns a matrix of given rank which approximates (in terms of SVD) given matrix m
}
