% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_utils.R
\name{dimensionality_reduction}
\alias{dimensionality_reduction}
\title{SVD based dimensionality reduction}
\usage{
dimensionality_reduction(m, dims, projection = FALSE)
}
\arguments{
\item{m}{given matrix to be projected}

\item{dims}{dimensionality of resulting matrix}

\item{projection}{project resulting matrix to dims-1 dimensional hyperplane}
}
\description{
Projection of given amtrix m to space of rank dims
}
\details{
If needed given points are projected to hyperplane in resulting space
}
