% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{evalClusters}
\alias{evalClusters}
\title{Evaluate accuracy of combination}
\usage{
evalClusters(dataset, clustering, cls, clusterNumber, cellTypesNumber)
}
\arguments{
\item{dataset}{matrix, given dataset to perform deconvolution}

\item{clustering}{numeric vector, cluster of every row in dataset}

\item{cls}{numeric vector, given cluster to evaluate accuracy}

\item{clusterNumber}{numeric, total number of clusters in dataset}

\item{cellTypesNumber}{numeric, expected number of cell types}
}
\value{
vector of cluster, sum-to-one error and log frobenius norm
}
\description{
Uses given combination as putative signatures for deconvolution
Runs DSA algorithm
}
\details{
Then calculates Frobenius norm between logarithms of
observed gene expression matrix
and estimated gene expression matrix
}
