% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerRanks.R
\name{geneMarkerStats}
\alias{geneMarkerStats}
\title{Get gene stats}
\usage{
geneMarkerStats(dataset, endpoints)
}
\arguments{
\item{dataset}{gene expression dataset}

\item{endpoints}{signature gene expression patterns}
}
\value{
data.frame with different statistics for genes in dataset
}
\description{
Gets gene statistics corresponding to found simplex ndpoints
}
