% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapClusters.R
\name{heatmapClusters}
\alias{heatmapClusters}
\title{Draw a heatmap of clustered dataset}
\usage{
heatmapClusters(dataset)
}
\arguments{
\item{dataset}{matrix or data.frame of clustered dataset or path to file}
}
\description{
Draws a heatmap of dataset clustered by preprocessDataset
If pheatmap is presented, heatmapClusters uses it
Othwerise stats::heatmap is used
}
\examples{
\dontrun{
clusteredDataset <- preprocessDataset("GSE19830")
heatmapClusters(clusteredDataset)
heatmapClusters("/path/To/Dataset.tsv")
}
}

