% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countInside.R
\name{hyperVolume}
\alias{hyperVolume}
\title{Compute the volume of the given simplex}
\usage{
hyperVolume(V)
}
\arguments{
\item{V}{matrix, given simplex, n \times (n + 1) matrix describing (n + 1) points in n-dimensional space}
}
\value{
float, hypervolume of given simplex
}
\description{
Compute the volume of the given simplex
}
