% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseLinearFit.R
\name{pairwiseDemingRegression}
\alias{pairwiseDemingRegression}
\title{Pairwise Deming regression fits between genes}
\usage{
pairwiseDemingRegression(X)
}
\arguments{
\item{X}{matrix, given gene expression dataset, rows supposed to be genes}
}
\value{
list containing two N x N matrices: R2 and slopes
}
\description{
Performs pairwise linear fit gene1 ~ k * gene2
}

