% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProportions.R
\name{plotProportions}
\alias{plotProportions}
\title{Draw a plot of estimated proportions}
\usage{
plotProportions(..., pnames = NULL)
}
\arguments{
\item{proportions}{matricies, data frames or NMF objects of estimated proportions or paths to file}
}
\description{
Draws a plot of estimated proprotions
If ggplot2 and reshape2 are installed will use them and return ggplot object
Otherwise will use standart R functions
}
\examples{
\dontrun{
# lets assume we already have clustered dataset and estimated accuracy
deconvolutionResults <- chooseBest(dataset, accuracy)
plotProportions(deconvolutionResults)
plotProportions(coef(deconvolutionResults)) # same as previous call
plotProportions("/path/to/deconvolution_results.tsv")
}
}

