% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_utils.R
\name{point_selection}
\alias{point_selection}
\title{Selection of points at the ends of simplex using VCA algorithms several times}
\usage{
point_selection(reduced, dims, number_of_iterations = 100)
}
\arguments{
\item{reduced}{matrix or data.frame object representing objects in reduced space}

\item{dims}{positive number, dimensionality of reduced space}

\item{number_of_iterations}{number of iterations to perform VCA}
}
\value{
data.frame contains endpoints of reduced dataset and associated cluster number
}
\description{
Selection of points at the ends of simplex using VCA algorithms several times
}
