% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_utils.R
\name{point_selection_iea}
\alias{point_selection_iea}
\title{Selection of points at the ends of simplex using IEA algorithms several times}
\usage{
point_selection_iea(x, dims, number_of_iterations = 100, toNormalise = T)
}
\arguments{
\item{dims}{positive number, dimensionality of reduced space}

\item{number_of_iterations}{number of iterations to perform VCA}

\item{reduced}{matrix or data.frame object representing objects in reduced space}
}
\value{
data.frame contains endpoints of reduced dataset and associated cluster number
}
\description{
Selection of points at the ends of simplex using IEA algorithms several times
}
