% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDSA.R
\name{runDSA}
\alias{runDSA}
\title{Run DSA}
\usage{
runDSA(dataset, clustering, clusters)
}
\arguments{
\item{dataset}{gene expression matrix}

\item{clustering}{numeric vector, clustering of the rows}

\item{clusters}{numeric vector, which clusters use as putative signatures}
}
\value{
NMF object, deconvolution results
}
\description{
Runs DSA with provided clusters as putative signatures
}

