% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countInside.R
\name{simplexDensity}
\alias{simplexDensity}
\title{Check how many points are inside of given simplex}
\usage{
simplexDensity(V, P)
}
\arguments{
\item{V}{matrix, given simplex, n \times (n + 1) matrix describing (n + 1) points in n-dimensional space}

\item{P}{matrix, given points to check, n \times N matrix describing N points in n-dimensional space}
}
\value{
float, how "dense" points are in given simplex
}
\description{
Number of points inside of given simplex is devided by hypervolume of given simplex
}
