% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{chooseBest}
\alias{chooseBest}
\title{Choose best combination of clusters and deconvolve}
\usage{
chooseBest(dataset, accuracy)
}
\arguments{
\item{dataset}{matrix, given dataset to perform deconvolution}

\item{accuracy}{matrix, given accuracy table, result of clusdecAccuracy function}
}
\value{
list with matrices W and H
}
\description{
Chooses best combination of clusters in accuracy table and then
runs DSA using this combination of clusters as putative signatures
}
\examples{
data('datasetLiverBrainLung')
preprocessed <- preprocessDataset(datasetLiverBrainLung, k=5) # 5 clusters
accuracy <- clusdecAccuracy(preprocessed, 3) # assuming 3 cell types
results <- chooseBest(preprocessed, accuracy) # choose best combination of clusters as putative sigantures


}
