% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{clusdecAccuracy}
\alias{clusdecAccuracy}
\title{Accuracy for all combination of clusters}
\usage{
clusdecAccuracy(dataset, cellTypesNumber, cores = 1)
}
\arguments{
\item{dataset}{matrix, clustered dataset}

\item{cellTypesNumber}{numeric, expected number of cell types}

\item{cores}{numeric, how many cores to use}
}
\value{
matrix, matrix of estimated accuracy (log Frobenius norm) for every combination
}
\description{
Accuracy for all combination of clusters
}
\examples{
data('datasetLiverBrainLung')
preprocessed <- preprocessDataset(datasetLiverBrainLung, k=5) # 5 clusters
accuracy <- clusdecAccuracy(preprocessed, 3) # assuming 3 cell types

\dontrun{
accuracy <- clusdecAccuracy(preprocessed, 3, cores=2) # using 2 CPU cores
}

}

