% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{deconvolveClusters}
\alias{deconvolveClusters}
\title{Deconvolution by given set of clusters}
\usage{
deconvolveClusters(dataset, clusters)
}
\arguments{
\item{dataset}{matrix, given dataset to perform deconvolution}

\item{clusters}{numeric vector, clusters to use as putative signatures}
}
\value{
list with matrices W and H
}
\description{
Runs DSA with  given clusters as putative signatures
}

